Version 1
First draft of the website. Day 1.

Need to look into an issue with the 'mouseover' event that causes the mouse to not register on a tile when the mouse if moving too fast.

Also need to fix the random colour button and it working on the tiles. I can get it working using .style, but I need to see if there's a way to use the random colour function whilst using classes to alter the tile colours.